refactor: Move on_shutdown to a separate module
