+++ bot/utils/on_shutdown.py
@@ -1,0 +1,12 @@
+import logging as log
+
+from aiogram import Dispatcher
+
+
+async def on_shutdown(dp: Dispatcher):
+    await dp.bot.delete_webhook()
+
+    await dp.storage.close()
+    await dp.storage.wait_closed()
+
+    log.warning("BOT STOPPED!")
